Разделяет JS-код на модули
